Fixed the particle spawner ids u32 overflow, call the CSM's core.register_on_sending_chat_message correctly

+++ src/client/core/clientevent.h
@@ -95,7 +95,7 @@ struct ClientEvent
 		{
 			ParticleSpawnerParameters *p;
 			u16 attached_id;
-			u64 id;
+			u32 id;
 		} add_particlespawner;
 		struct
 		{
